+++ Makefile
@@ -11,13 +11,13 @@ debug: CFLAGS += -coverage -g -fsanitize=undefined
 debug: all libFJML.so
 	sudo make install
 
-HEADERS = src/FJML/activations/activations.h \
-		  src/FJML/data/data.h \
-		  src/FJML/layers/layers.h \
-		  src/FJML/linalg/linalg.h src/FJML/linalg/tensor.h \
-		  src/FJML/loss/loss.h \
-		  src/FJML/mlp/mlp.h \
-		  src/FJML/optimizers/optimizers.h
+HEADERS = include/FJML/activations.h \
+		  include/FJML/data.h \
+		  include/FJML/layers.h \
+		  include/FJML/tensor.h \
+		  include/FJML/loss.h \
+		  include/FJML/mlp.h \
+		  include/FJML/optimizers.h
 CFILES = bin/activations/activations.o \
 		 bin/layers/dense.o bin/layers/layers.o bin/layers/softmax.o \
 		 bin/loss/loss.o \
@@ -32,15 +32,14 @@ all: $(CFILES)
 
 install: libFJML.so
 	rm -rf /usr/local/include/FJML
-	cp -r src/FJML /usr/include
-	cp src/FJML.h /usr/include
+	cp -r include/* /usr/local/include
 	cp libFJML.so /usr/local/lib
 	ldconfig /usr/local/lib
 
 init:
 	mkdir -p bin/layers bin/activations bin/loss bin/util bin/mlp bin/optimizers
 
-docs: src/FJML/**/* doxygen.conf
+docs: src/**/* include/**/* doxygen.conf
 	doxygen doxygen.conf
 
 coverage:
