+++ include/FJML.h
@@ -1,0 +1,13 @@
+#ifndef FJML_INCLUDED
+#define FJML_INCLUDED
+
+#include "./FJML/activations.h"
+#include "./FJML/data.h"
+#include "./FJML/layers.h"
+#include "./FJML/linalg.h"
+#include "./FJML/tensor.h"
+#include "./FJML/loss.h"
+#include "./FJML/mlp.h"
+#include "./FJML/optimizers.h"
+
+#endif
