--- src/FJML.h
@@ -1,13 +1,0 @@
-#ifndef FJML_INCLUDED
-#define FJML_INCLUDED
-
-#include "./FJML/activations/activations.h"
-#include "./FJML/data/data.h"
-#include "./FJML/layers/layers.h"
-#include "./FJML/linalg/linalg.h"
-#include "./FJML/linalg/tensor.h"
-#include "./FJML/loss/loss.h"
-#include "./FJML/mlp/mlp.h"
-#include "./FJML/optimizers/optimizers.h"
-
-#endif
