+++ src/FJML/layers/layers.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) 2023 David Lee
 // This code is licensed under MIT license (see LICENSE for details)
 
-#include "layers.h"
+#include "../../../include/FJML/layers.h"
 
 namespace FJML {
